Specify a character for escaped substitution
When the name of the variable starts with a character that is normally interpreted within
mustache templating (i.e. "#^/{&!>:") it should be possible to use an escape character.
This commit propose to use colon (:) as first character to escape the first character of
variable name in the variable substitution with escaping.
Writing {{:NAME}} is equivalent to writing {{NAME}} even when
NAME begins with a special character.

It is interesting to see that it specify an optional character for variable
substitution with escaping and not only an escaping character.

Signed-off-by: José Bollo <jose.bollo@iot.bzh>